Move type extensions to provider because otherwise they don't seem to load
